requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
openai>=1.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
httpx>=0.24.0

